requirements.txt
datasets-plus
json-repair
litellm
tqdm
matplotlib
seaborn
sentence-transformers
loguru